Recompiled the heatmap_gen for Windows
The executable for the Windows version of heatmap_gen is located in HeatmapBuild/Debug/HeatmapBuild.exe. Since the Linux version used cout for the newlines, Linux and Windows interpreted the newlines differently so the Windows compilation had some extra data which corrupted the png file created as well as some missing libraries. _setmode(_fileno(stdout), _O_BINARY); added to the heatmap_gen.cpp main was able to fix the machine specific differences.