Fixed heatmap executable and execution in Unity
The debug version of the executable did not work with the Unity Editor, so I recompiled it in a release version for other programs to use. I also changed the way that Unity executes the file--instead of directly calling it, Unity executes it through opening cmd.exe on Windows. The heatmaps are able to be generated, but they aren't able to be parsed at runtime just yet.